bug-1911612: move logger decorator to socorro/lib/util.py
This decorator will ultimately be used by both the ESCrashStorage class and the SuperSearch class, so it makes sense to put it somewhere more central.
